feat: Removed callkeep, added notification
Callkeep is not behaving as expected, so using flutter local
notification as a makeshift solution.
With callkeep plugin, the caller UI freezes when the call is accepted.
From the logs it seems like the call is placed more than once on the
plugin side, and hence the issue. Needs further investigation